Linked Inventory and ItemEdit together
