+++ client/src/views/Inventory.vue
@@ -1,42 +1,39 @@
 <template>
-<div class="inventory">
-  <AdminHeader/>
-  <body>
-    <h2 class="title is-2">Add/Edit Item</h2>
-    <input class="input" type="text" placeholder="Item Name">
-    <br>
-    <input class="input" type="text" placeholder="Image URL">
-    <br>
-    <input class="input" type="text" placeholder="Price">
-    <br>
-    <input class="input" type="text" placeholder="Quantity">
-    <br>
-    <br>
-
-    <div class="control">
-      <label class="radio">
-        <input type="radio" name="answer">
-        In Store Only
-      </label>
-      <label class="radio">
-        <input type="radio" name="answer">
-        Can Ship
-      </label>
-      <br>
-      <br>
-      <button class="button" style="margin-right: 15px">
-        <router-link to="/inventory">Save</router-link>
-      </button>
+  <div class="inventory">
+    <AdminHeader/>
+    <h2 class="title is-2">Inventory</h2>
+    <div>
       <button class="button">
-        <router-link to="/inventory">Cancel</router-link>
+        <router-link to="/itemedit" exact-active-class="is-active">Add Item</router-link>
       </button>
     </div>
-  </body>
-</div>
+    <div v-for="(Part, index) in myParts" v-bind:key="index">
+      <div class="table">
+        <div class="row">
+          <div class="cell col1">
+            <input type="checkbox" name="name1" id="1">
+            <br>
+          </div>
+          <div class="cell col2">
+            <span>{{ Part.name }}</span>
+          </div>
+          <div class="cell col4">
+            <button class="button" name="btn-delete">Delete</button>
+            <button class="button" name="btn-Edit">
+              <router-link to="/itemedit" exact-active-class="is-active">Edit</router-link>
+            </button>
+          </div>
+        </div>
+      </div>
+    </div>
+    <br>
+  </div>
 </template>
 
-<script lang="ts">
-import { Component, Vue } from "vue-property-decorator";
+<script lang='ts'>
+import Vue from "vue";
+
+import { Component } from "vue-property-decorator";
 import AdminHeader from "@/components/AdminHeader.vue";
 
 @Component({
@@ -44,7 +41,21 @@ import AdminHeader from "@/components/AdminHeader.vue";
     AdminHeader
   }
 })
-export default class Inventory extends Vue {}
+export default class Inventory extends Vue {
+  // this export says, if another files imports this file,
+  // then it only gets this component
+  myParts: Part[] = [
+    { name: "Trek Bike" },
+    { name: "Bike Pump" },
+    { name: "Tire" }
+  ];
+  item: Part = {
+    name: ""
+  };
+}
+interface Part {
+  name: string;
+}
 </script>
 
 <style scoped>
